# RFC: eliminate wasm_rt_register_func_type in wasm2c runtime API