… CMakeLists.txt
… test/wasm2c/add.txt
… test/wasm2c/hello.txt
… test/wasm2c/minimal.txt
… wasm2c/README.md
… wasm2c/examples/fac/fac.c