Kicked off by #2117, here is a PR that demonstrates how we could eliminate the `wasm_rt_register_func_type` API call (and the need to worry about synchronizing runtime-wide state for the function type registry).

Previously, wasm2c-generated modules registered function types with the runtime at module initialization time. Two identical function type signatures are required to receive the same global index (this is how call_indirect checks type equality at runtime).

With this PR, wasm2c generates an identifier that serializes each function type at "transpile-time." The serialized function type is either:
- a uint64_t (if the function has $\le$ 16 params + results), or
- a length-tagged string

This lets us eliminate the need to register function types at module-initialization time.

I did a bunch of microbenchmarking of the performance impact by running a billion iterations of a foreign-module call_indirect (benchmark here: https://gist.github.com/keithw/aaf17bf55165fe0de522c5c8dbfc6d4d). I ran the benchmark on an Intel i7-9700T (burstable to 4.3 GHz single-core) and an AMD EPYC 7702 (burstable to 3.35 GHz single-core) with gcc 12 and clang 15, also trying out different `-march` settings (which had a substantial effect). I ran each benchmark three times and averaged the results; here are the times per-iteration.

| | Intel gcc | Intel gcc `-march=znver2` | Intel clang | AMD gcc | AMD clang |
|------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| main branch | 1.6 ns | 1.6 ns | 1.4 ns |1.5 ns | 1.5 ns|
| this PR | 1.4 ns | 1.2 ns | 1.9 ns  |2.1 ns |2.1 ns |

At this level I think we can persuade ourselves that the performance impact is pretty small and maybe is measuring vagaries of alignment of data structures as much as anything else. The results varied between a 30% speedup to a 40% slowdown, but were always within roughly one CPU cycle up or down. (It would probably be worse for types with > 16 params/results.)

On the broader design questions, I think I've gotten myself to a point where the approach here seems like the least-bad option. Here are some possible alternatives:

1. Keep the status quo, but let the runtime provide a thread-safe `wasm_rt_register_func_type` by locking it internally (#2117). I don't love requiring the runtime to maintain this eternally growing global registry, with no easy way to garbage-collect, and the need to synchronize access, if we can avoid it. If we want the runtime to be re-entrant, I am optimistic that we can eliminate basically all the need for synchronized global state.
2. Make the runtime allocate instances of the registry on the heap and return an opaque context object, and require the application to feed this to modules at their initialization time. I thought this would work, but the problem is then... where does the module store the association between its type indexes and the runtime-wide type addresses?
   - If the module this this in its own static `func_types` array (like today), then... is that module bound to only work with one runtime instance (and other modules initialized with the same runtime instance), and the user has to keep track of all this? That seems fragile and likely to break.
   - Alternately, the module could put its func_types mapping in a new "module state" structure, and then that structure would probably have to be passed in as an additional parameter to every call of a Wasm function, along with the module instance pointer. This seems pretty gross ergonomics.
   - Alternately, instead of passing it as a function parameter, the "module state" structure could be pointed-to by every module instance. But then we'd have different "colored" instances based on which runtime instance they were associated with, and the user would have to be careful not to instantiate a module instance given an exporting module instance that came from a different runtime instance? That seems even worse ergonomics.
   
All things considered, then, maybe this is the least-bad way to make everybody happy long-term. Open to thoughts though...