@keithw 2023-01-06
Closing as I think there may be a much simpler way to make everybody happy here.