@keithw 2023-01-07
And, reopening after some the "much simpler way" turned out to be more vexing than I had hoped (see comment).