@shravanrn 2023-01-07
@keithw The broad approach looks good to me. However, I think the one change that would be great is if we can pull the type serializer into its own separate file/library so I can pull this in for RLBox also. This would let me avoid re-implementing the type serializer in RLBox to dynamically construct a type index. 

One point I had when I quickly look at the code is that the complexity of the type serializer makes me a bit nervous, however I don't know of a way to make this any simpler. I was hoping to do a closer code review, but I am traveling at the moment, so I may only be able to get to this in a few days.