@keithw 2023-01-11
> However, I think the one change that would be great is if we can pull the type serializer into its own separate file/library so I can pull this in for RLBox also.

I think that would be quite doable, but there's a question about how you want to use it. E.g. are you able to link with libwabt to get it? That would be one easy way to make this happen without duplicating anything.