@keithw 2023-01-13
Or... maybe the easiest approach here is for wasm2c to stick a bunch of `#define`s in the generated header file, giving a reasonable name for each type in the module, e.g.:
```c
#ifndef Z_w2c_type_param_i32_i64_return_externref
#define Z_w2c_type_param_i32_i64_return_externref (wasm_rt_func_type_t){{.val = 260}, 8})
#endif
```

That way you could just use the type directly in your code (worst case the type doesn't exist in any module and the code fails to compile, which should be an error anyway since there's no way it can successfully be call_indirect'ed), and it doesn't seem like _that_ big of a deal to pollute the global (preprocessor) namespace with these `#define` names.