@keithw 2023-01-17
I went through a bunch of revisions on this and I think I'm happy where it's landed now. Because of string interning/constant pooling in the GNU/LLVM toolchains, there's essentially no performance impact (no matter the size of the function signature), even when making a call_indirect across modules.

In cases where the actual 256-bit values have to be compared (i.e., the call_indirect crosses modules AND the toolchain didn't merge the string constants across compilation units), it's generally a 1 or 1.5 cycle penalty.