@shravanrn 2023-01-18
Sorry catching up after 10 days away

> I think that would be quite doable, but there's a question about how you want to use it. E.g. are you able to link with libwabt to get it? That would be one easy way to make this happen without duplicating anything.

It will be a tough sell to link all of wabt in Firefox, due to the binary bloat. Realistically, this would need to be separated out into its own files. (I am happy to help pull this out if needed). 

That said, I don't think this needs to change the internal library configuration in the wabt project. If we can pull out just the function that generates the func type into its own file which only imports wasm-rt.h and standard C headers, this would be perfectly fine. I am happy to pull just that one file into the runtime we would use for Firefox.

> Other option would be to just duplicate the functionality in the wasm runtime 

Do you mean in the runtime that is checked in? Sure, that would be fine. If you mean duplicate it in the rlbox-wasm runtime, that would be fine if I can easily pull in just these files and not all of wabt (same as prev point). The only thing i'd want to avoid is having a second implementation as it feels like things can get out of sync. 

> Or... maybe the easiest approach here is for wasm2c to stick a bunch of #defines in the generated header file, giving a reasonable name for each type in the module, e.g.:

Unfortunately, this would not be usable by RLBox. RLBox tries to make dealing with wasm'd libraries identical to using non-Wasm'd libraries. Thus when handling callbacks (host exposed functions), RLBox automates things like func_type index lookup by (1) reflecting on the callback function signature and (2) calling the procedure to get the func_type index based on the reflection results. Unfortunately, C++ does not offer a way to pick one of many macros based on reflection. Thus a simple static procedure for func type index generation `u32 get_func_idx(wasm_rt_type_t  ret, wasm_rt_type_t params ... )` would be preferable. 

I'll go through the latest changes to the PR shortly. Some quick thoughts 
- Using strings for comparison and relying on merging for perf is clever.
- I didn't fully get where the hashing is used yet, but I would love to avoid relying on an extra dependency just for this.

I'll have a closer look at the code and get back