@shravanrn 2023-01-18
@keithw Just a thought: Could we skip the hashing step altogether and simply output the mangled string? That way, we cam compare with `strcmp`. Since the string merging will most likely reduce this to a pointer compare, it seems like we would maintain the perf benefits without pulling in a dependency for hashing.