@keithw 2023-01-19
> @keithw Basic question: Could we skip the hashing step altogether and simply output the mangled string? That way, we can compare with `strcmp`. Since the string merging will most likely reduce this to a pointer compare, it seems like we would maintain the perf benefits without pulling in a dependency for hashing.

I guess I'd love to avoid thinking about the uncertain cost of comparing unbounded-length strings at runtime. (The original version of this PR of course had that issue, but worked hard to fit "most" function types in a single uint64_t -- at the cost of a somewhat complex serializer that is gone now.)

It seems comforting to know that the worst case is a single 256-bit compare (which is a single cycle when compiling for contemporary x86-64 with AVX-2, and about 1.5 cycles on x86-64 otherwise). Of course all this only applies when the linker wasn't smart enough to merge the string constants in the first place (but I'm not sure how reliable that happens for arbitrary-length string constants -- again it seems more comfortable to know they're always 32 bytes exactly).

> EDIT: just saw the concern about cross module perf. If this is a concern and we want to stick with hashing, I think PicoSHA2 maybe a nicer option, just so that wabt runtime doesn't pull in more external dependencies than necessary (in essence, avoiding all parts of openssl would be my preferred option :) )

Well, not sure if this was clear, but the wasm2c **runtime** has zero added dependencies from this. No hashing happens at runtime -- only at wasm2c-time.

The wasm2c tool does end up needing to do SHA-256 somehow. In the current version of this PR, CMake tries to find OpenSSL, but if it's configured not to via the command line (or if it's not available in the places that CMake looks), wasm2c ends up using the vendored (internal) PicoSHA2 and the resulting wasm2c doesn't end up with any external library dependency. So if you want a wasm2c free of OpenSSL dependencies, that's pretty easy to arrange (and either way the runtime and compiled code have no hashing at all, whether OpenSSL or internal). How do you feel about that / does it work for you?