@deian 2023-01-19
Cool stuff. Personally I think the other option is safer and simpler.

Bringing OpenSSL in the TCB seems super scary and for a bunch of use cases would conflict with system that ship with crypto libs (NSS, libsodium, etc.). It's also not obvious what the performance implications are beyond these micro benchmarks (in contrast, the simpler lock approach already met demands of workloads where latency matters).